+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_hair_table.hpp
@@ -6,9 +6,9 @@
 class crhair
 {
   public:
-    uint8_t h_hair;
+    uint64_t h_hair;
 public:
-    uint8_t primary_key() const {return h_hair;}
+    uint64_t primary_key() const {return h_hair;}
 };
 
 typedef multi_index<N(crhair),crhair> hair_data_table;
