+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_head_table.hpp
@@ -6,9 +6,10 @@
 class crhead
 {
   public:
-    uint8_t h_head;
-public:
-    uint8_t primary_key() const {return h_head;}
+    uint64_t h_head;
+
+  public:
+    uint64_t primary_key() const { return h_head; }
 };
 
 typedef multi_index<N(crhead),crhead> head_data_table;
